sks ta: CKM_RSA_PKCS relies on TEE_ALG_RSASSA_PKCS1_V1_5
TEE_ALG_RSASSA_PKCS1_V1_5 is not part of the GPD TEE standard
and was recently added as an OP-TEE extension of the Core Internal
API. CKM_RSA_PKCS mechanism signing and verifying relies on this scheme
were the payload hash is computed and supplied by the client.

The TEE_ALG_RSASSA_PKCS1_V1_5 currently uses EMSA padding and
signs the message without hashing it.

Signed-off-by: Gabor Szekely <szvgabor@gmail.com>
[Rebase on top of master and split commit in smaller changes]
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>