stm32mp1: more flexible UART console
Add a timeout in output console. This is useful if the secure world
relies on a non secure UART that may be suspended or clock down by
the non secure world.

Rename structure console_pdata into stm32_uart_pdata as it will
be exported over the platform.

Prepare to map all non secure UARTs and the shared UART which could be
secure or non secure.

CFG_STM32_EARLY_CONSOLE_UART also to specify the hard coded UART
interface used for early trace console.

Signed-off-by: Etienne Carriere <etienne.carriere@st.com>