[DO NOT MERGE] core: pseudo TA can produce a session id
Remove opaque handle user_ctx from struct tee_ta_session that
is used by pseudo TAs to safely store the session ID returned
to client at session opening. It is replaced with a session
list, tracked by the pseudo TA.

It is not worth it! This change removes a 32 or 64bit cell from
all session instances at the cost of a list to manage. All
pseudo TAs shared to same privileged execution context and
session are already tracked with lists and context handles. No
need to add extra complexity.

This is why you shall reject this P-R.

Signed-off-by: Etienne Carriere <etienne.carriere@st.com>