stm32mp1: shres: register always secure clocks
Some platform clocks are always secure. Such are not listed
in the shared resources since they cannot be assigned to the
non-secure world. However such clock may be generated by a
parent clock which is such case should be assigned to the
secure world. This change register registers as secure the
parent clocks of the always secure clocks.

Signed-off-by: Etienne Carriere <etienne.carriere@st.com>