ta: sks: change cmd handlers to check param types
Param#0 is now a mandatory in/out memref parameter. It is used as an
input buffer for almost all commands as almost all commands expected
specific structured arguments. It is used by all commands as an
output buffer for the TA to report to client a PKCS#11 compliant
status for the invoked command.

As a consuquence, TA commands that do not expect arguments from
param#0 only check that param#0 buffer is 32bit large to store
the 32bit PKCS#11 compliant status.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>