ta: sks: set default token serial number based on token id
It is common for the client application to decide the slot to use
based on the token serial number. Therefore change the default value
to be based on the token ID to avoid having the same serial numbers
on every token. This change was suggested by [1].

With this change, token serial number is string "0...0000000<N>" over
16 characters, led with '0' and ending with token ID decimal value.

Link: [1] https://github.com/ricardosalveti/optee-sks/commit/b82604d4a3172d3ac2eae4ff7938d64fd5b25432
Suggested-by: Ricardo Salveti <ricardo@foundries.io>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>