stpmic1: relax helper function to not panic on invalid regulator IDs
Change STPMIC1 driver local helper function get_regulator_data() to
not panic but return a NULL reference when provided regulator ID
is invalid. No functional change aside an explicit panic() being
replaced with a segmentation fault when a invalid reference is used.

This change allows caller to decide whether to panic or report an error
when an invalid STPMIC1 regulator ID is used.

Change-Id: I5af5ef9b0f7cd18bf563f1de56de26a695cdd6a4
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>