ta: sks: add a default empty value for expected object attributes
When sanitizing a template the create a object, set missing optional
expected object attributes to a empty value. This change syncs with
OP-TEE mainline implementation as of 3.10.0.

As a consequence, such empty attributes must be removed before
the effective attribute is computed and added in related object
processing as object creation, generation or derivation.

As a side effect, add PKCS11_CKA_EC_POINT as EC private key optional
attribute since this attribute is expected empty in application
key manipulation sequences.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>