scmi-server: pta command GET_CHANNEL without shm info
Support new ABI for get_channel: no static shared memory location.
When client do not provide a SHM reference in GET_CHANNEL, it is
expected to register a shm later and use it to communicate with
the server.

This change support old and new GET_CHANNEL. Actually, the new command
(without shm ref) is quite useless since nothing allows to identify
the caller agent, but the agent-id it already provides hence this change
makes SCMI PTA return this agent ID to the agent/client as channel ID.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>