ta: pkcs11: Load persistent object attributes during db initialization
Persistent object attributes are now loaded during the first find
operation. This means that you're forced to do a find operation to make
the objects available.

Right now entry_find_objects_init is broken in such that it tries to
access the obj->attributes for objects in the session->token->object_list
before setting the obj->attributes (by calling token_obj_matches_ref).
This results in a Panic.

This patch fixes both issues.

Signed-off-by: Robin van der Gracht <robin@protonic.nl>