core: arm: mobj: some mobjs may have no physical address
Change mobj_with_fobj_get_pa() to return TEE_ERROR_NO_DATA
when a virtual memory address has no assigned physical address
which is the case when the related memory is pageable and pager
is enabled. This is the only memory object (mobj) for which the
object physical address range is volatile because under pager control.

Update vm_pa2va() to not report a failure in such case but just skip
the memory region.

Update check_pa_matches_va() to not panic when vm_va2pa() returns this
code because it can't convert the virtual address because the effective
physical address of the memory cell is volatile as when target memory
is paged and pager is enabled.

All other calls to mobj_get_pa() already handle the return code values
they need to.

For consistency this changes add an inline description to mobj_get_pa()
to describe the possible return values of the helper function.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>